+++ src/interface/message.rs
@@ -14,7 +14,10 @@ use crate::usecase::Request;
 #[derive(Debug)]
 pub enum Message {
     Teams(Vec<Team>),
-    Viewer(UserId),
+    Viewer {
+        id: UserId,
+        organization: Option<Organization>,
+    },
     TeamContext {
         team_id: TeamId,
         states: Vec<WorkflowState>,